invariant-violation -> asserts
